added more possible option lists and cleaned up some code by removing unneccsary bootstrap files

+++ application/src/App.js
@@ -15,17 +15,12 @@ class App extends Component {
   constructor (props) {
       super(props);
       this.state = {
-        modal: false,
-        chosen: null,
-        buttonHistory: [],
-        currentButton: 0,
-
-        questions:{ }
+        modal: false
       };
 
   this.toggle = this.toggle.bind(this);
-this.button_chosen = this.button_chosen.bind(this);
-this.SetStart = this.SetStart.bind(this)
+
+
 
   }
   toggle() {
@@ -33,23 +28,6 @@ this.SetStart = this.SetStart.bind(this)
       modal: !this.state.modal
     });
   }
-  SetStart = () => {
-    this.setState({chosen: null});
-  }
-  button_chosen = (position,q)=>{
-    if(this.state.chosen === position){
-      this.setState({chosen: null})
-    }
-    else{
-      this.setState({chosen:position})
-    }
-
-  }
-
-
-button_page_check = (positon, q) => {
-  if (q.questionid ){}
-}
 
 
 traverser(){
@@ -272,6 +250,26 @@ else if(type === 1){
   sessionStorage.setItem(q.questionid, '1');
 
 }
+else if(type === 2){
+  sessionStorage.setItem(q.questionid, '2');
+
+}
+else if(type === 3){
+  sessionStorage.setItem(q.questionid, '3');
+
+}
+else if(type === 4){
+  sessionStorage.setItem(q.questionid, '4');
+
+}
+else if(type === 5){
+  sessionStorage.setItem(q.questionid, '5');
+
+}
+else if(type === 6){
+  sessionStorage.setItem(q.questionid, '6');
+
+}
 
 }
 chosen_color_0 = (q) =>{
@@ -291,7 +289,46 @@ if(sessionStorage.getItem(q.questionid)=== '1'){
   else{
   return "";}
 }
+chosen_color_2 = (q) =>{
+if(sessionStorage.getItem(q.questionid)=== '2'){
+    return "blue";
 
+  }
+  else{
+  return "";}
+}
+chosen_color_3 = (q) =>{
+if(sessionStorage.getItem(q.questionid)=== '3'){
+    return "blue";
+
+  }
+  else{
+  return "";}
+}
+chosen_color_4 = (q) =>{
+if(sessionStorage.getItem(q.questionid)=== '4'){
+    return "blue";
+
+  }
+  else{
+  return "";}
+}
+chosen_color_5 = (q) =>{
+if(sessionStorage.getItem(q.questionid)=== '5'){
+    return "blue";
+
+  }
+  else{
+  return "";}
+}
+chosen_color_6 = (q) =>{
+if(sessionStorage.getItem(q.questionid)=== '6'){
+    return "blue";
+
+  }
+  else{
+  return "";}
+}
 
 clear_storage= () =>{
   sessionStorage.clear();
@@ -307,7 +344,7 @@ clear_storage= () =>{
 <Route>
 <div >
 <h1 className = "titlebg" >
-                 <img className = "img" src="https://osulibrary.oregonstate.edu/sites/all/themes/doug-fir-d7-library/logo.svg" alt="osu" width="100" height="100"></img> DC Wizard <NavLink to="/"><Button className = "Restart" onClick={()=>{this.SetStart; this.clear_storage()}}>Restart</Button></NavLink></h1>
+                 <img className = "img" src="https://osulibrary.oregonstate.edu/sites/all/themes/doug-fir-d7-library/logo.svg" alt="osu" width="100" height="100"></img> DC Wizard <NavLink to="/"><Button className = "Restart" onClick={()=>{ this.clear_storage()}}>Restart</Button></NavLink></h1>
 
 
          <ul  className="header">
@@ -327,6 +364,7 @@ clear_storage= () =>{
 
          {
          data.map((q) =>  {
+
            while(q.questionid == window.location.pathname){
             if(q.questionid.indexOf('done') >= 0){
               return<div>
@@ -339,6 +377,7 @@ clear_storage= () =>{
             }
             else{
               if(q.numoptions == 2){
+
              return<div>
 
              <h1>Printing out from Json: {q.question}</h1>
@@ -396,12 +435,185 @@ clear_storage= () =>{
           <ul className="header">
 
 
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color_0(q)}} onClick={() => {this.question_show(q,0)}} >{q.optionid1.option}</Button></NavLink></li>
 
-          <li><NavLink to={q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color(0)}} onClick={() => {this.button_chosen(0)}} >{q.optionid1.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color_1(q)}} onClick={() => {this.question_show(q,1)}} >{q.optionid2.option}</Button></NavLink></li>
 
-          <li><NavLink to={q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color(1)}} onClick={() => {this.button_chosen(1)}} >{q.optionid2.option}</Button></NavLink></li>
 
-          <li><NavLink to={q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color(2)}} onClick={() => {this.button_chosen(1)}} >{q.optionid2.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color_2(q)}} onClick={() => {this.question_show(q,2)}} >{q.optionid3.option}</Button></NavLink></li>
+
+       </ul>
+     </div>
+
+        }
+
+        else if(q.numoptions == 4){
+
+          return<div>
+
+          <h1>Printing out from Json: {q.question}</h1>
+          <div className="navigationButtonsLeft">
+          <NavLink to={q.questionorigin}><Button >&lt; Back</Button></NavLink>
+
+             </div>
+          <Button onClick={this.toggle} >  <div className = "buttondiv">More Info</div></Button>
+          <p>  {q.explanationresources}</p>
+
+                <p>{q.explanationresources}</p>
+          <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
+            <ModalHeader toggle={this.toggle}>  Information </ModalHeader>
+            <ModalBody>
+         Print from json: {q.explanation}
+            </ModalBody>
+            <ModalFooter>
+
+              <Button color="primary" onClick={this.toggle}>Got it!</Button>{' '}
+              <Button color="secondary" onClick={this.toggle}>Cancel</Button>
+            </ModalFooter>
+          </Modal>
+
+          <ul className="header">
+
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color_0(q)}} onClick={() => {this.question_show(q,0)}} >{q.optionid1.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color_1(q)}} onClick={() => {this.question_show(q,1)}} >{q.optionid2.option}</Button></NavLink></li>
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color_2(q)}} onClick={() => {this.question_show(q,2)}} >{q.optionid3.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid4.nextstepcontent}><Button style={{background: this.chosen_color_3(q)}} onClick={() => {this.question_show(q,3)}} >{q.optionid4.option}</Button></NavLink></li>
+
+       </ul>
+     </div>
+
+        }
+        else if(q.numoptions == 5){
+
+          return<div>
+
+          <h1>Printing out from Json: {q.question}</h1>
+          <div className="navigationButtonsLeft">
+          <NavLink to={q.questionorigin}><Button >&lt; Back</Button></NavLink>
+
+             </div>
+          <Button onClick={this.toggle} >  <div className = "buttondiv">More Info</div></Button>
+          <p>  {q.explanationresources}</p>
+
+                <p>{q.explanationresources}</p>
+          <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
+            <ModalHeader toggle={this.toggle}>  Information </ModalHeader>
+            <ModalBody>
+         Print from json: {q.explanation}
+            </ModalBody>
+            <ModalFooter>
+
+              <Button color="primary" onClick={this.toggle}>Got it!</Button>{' '}
+              <Button color="secondary" onClick={this.toggle}>Cancel</Button>
+            </ModalFooter>
+          </Modal>
+
+          <ul className="header">
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color_0(q)}} onClick={() => {this.question_show(q,0)}} >{q.optionid1.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color_1(q)}} onClick={() => {this.question_show(q,1)}} >{q.optionid2.option}</Button></NavLink></li>
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color_2(q)}} onClick={() => {this.question_show(q,2)}} >{q.optionid3.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid4.nextstepcontent}><Button style={{background: this.chosen_color_3(q)}} onClick={() => {this.question_show(q,3)}} >{q.optionid4.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid5.nextstepcontent}><Button style={{background: this.chosen_color_4(q)}} onClick={() => {this.question_show(q,4)}} >{q.optionin5.option}</Button></NavLink></li>
+
+       </ul>
+     </div>
+
+        }
+        else if(q.numoptions == 6){
+
+          return<div>
+
+          <h1>Printing out from Json: {q.question}</h1>
+          <div className="navigationButtonsLeft">
+          <NavLink to={q.questionorigin}><Button >&lt; Back</Button></NavLink>
+
+             </div>
+          <Button onClick={this.toggle} >  <div className = "buttondiv">More Info</div></Button>
+          <p>  {q.explanationresources}</p>
+
+                <p>{q.explanationresources}</p>
+          <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
+            <ModalHeader toggle={this.toggle}>  Information </ModalHeader>
+            <ModalBody>
+         Print from json: {q.explanation}
+            </ModalBody>
+            <ModalFooter>
+
+              <Button color="primary" onClick={this.toggle}>Got it!</Button>{' '}
+              <Button color="secondary" onClick={this.toggle}>Cancel</Button>
+            </ModalFooter>
+          </Modal>
+
+          <ul className="header">
+
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color_0(q)}} onClick={() => {this.question_show(q,0)}} >{q.optionid1.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color_1(q)}} onClick={() => {this.question_show(q,1)}} >{q.optionid2.option}</Button></NavLink></li>
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color_2(q)}} onClick={() => {this.question_show(q,2)}} >{q.optionid3.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid4.nextstepcontent}><Button style={{background: this.chosen_color_3(q)}} onClick={() => {this.question_show(q,3)}} >{q.optionid4.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid5.nextstepcontent}><Button style={{background: this.chosen_color_4(q)}} onClick={() => {this.question_show(q,4)}} >{q.optionin5.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid6.nextstepcontent}><Button style={{background: this.chosen_color_5(q)}} onClick={() => {this.question_show(q,5)}} >{q.optionin6.option}</Button></NavLink></li>
+
+       </ul>
+     </div>
+
+        }
+        else if(q.numoptions == 7){
+
+          return<div>
+
+          <h1>Printing out from Json: {q.question}</h1>
+          <div className="navigationButtonsLeft">
+          <NavLink to={q.questionorigin}><Button >&lt; Back</Button></NavLink>
+
+             </div>
+          <Button onClick={this.toggle} >  <div className = "buttondiv">More Info</div></Button>
+          <p>  {q.explanationresources}</p>
+
+                <p>{q.explanationresources}</p>
+          <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
+            <ModalHeader toggle={this.toggle}>  Information </ModalHeader>
+            <ModalBody>
+         Print from json: {q.explanation}
+            </ModalBody>
+            <ModalFooter>
+
+              <Button color="primary" onClick={this.toggle}>Got it!</Button>{' '}
+              <Button color="secondary" onClick={this.toggle}>Cancel</Button>
+            </ModalFooter>
+          </Modal>
+
+          <ul className="header">
+
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid1.nextstepcontent}><Button style={{background: this.chosen_color_0(q)}} onClick={() => {this.question_show(q,0)}} >{q.optionid1.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid2.nextstepcontent}><Button style={{background: this.chosen_color_1(q)}} onClick={() => {this.question_show(q,1)}} >{q.optionid2.option}</Button></NavLink></li>
+
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid3.nextstepcontent}><Button style={{background: this.chosen_color_2(q)}} onClick={() => {this.question_show(q,2)}} >{q.optionid3.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid4.nextstepcontent}><Button style={{background: this.chosen_color_3(q)}} onClick={() => {this.question_show(q,3)}} >{q.optionid4.option}</Button></NavLink></li>
+
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid5.nextstepcontent}><Button style={{background: this.chosen_color_4(q)}} onClick={() => {this.question_show(q,4)}} >{q.optionin5.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid6.nextstepcontent}><Button style={{background: this.chosen_color_5(q)}} onClick={() => {this.question_show(q,5)}} >{q.optionin6.option}</Button></NavLink></li>
+          <li><NavLink to={process.env.PUBLIC_URL + q.optionid7.nextstepcontent}><Button style={{background: this.chosen_color_6(q)}} onClick={() => {this.question_show(q,6)}} >{q.optionin7.option}</Button></NavLink></li>
 
        </ul>
      </div>
